Team Color functionality
Changed the way the teams work. Now there are 2 team objects, each one has a name for displaying in the ui, a class for assigning to ui elements, and a color, which determines the color used for that team in all css styling.

Now we can very easily change the color and display names of the teams and eventually add UI functionality so that players can do this for themselves